+++ lib/src/node.dart
@@ -285,6 +285,7 @@ class Node<T> {
   // ...........................................................................
   void _initScm() {
     scm.addNode(this);
+    scm.initSuppliers(this);
     _dispose.add(() {
       scm.removeNode(this);